Series of updates
Remove Brewfile, put the brew formulas in the readme

Update the readme

Update the Rakefile

Some of the bash aliases work ok but the functions don't in Fish. These
will need to be modified.

Create a new fish config and force symlinking it into place on shell
start

Reorganize the vim plugins between critical ones and non-critical ones